Moved config module to src
