+++ src/bottles.py
@@ -3,7 +3,7 @@
 This includes all functions for the Lists, DB and Buttos/Dropdowns.
 """
 
-from config.config_manager import shared
+from src.config_manager import shared
 
 from src.database_commander import DB_COMMANDER
 from src.display_controller import DP_CONTROLLER
