+++ src/display_controller.py
@@ -4,7 +4,7 @@ from PyQt5.QtCore import Qt
 from src.database_commander import DB_COMMANDER
 from src.dialog_handler import DialogHandler, UI_LANGUAGE
 from src.models import Cocktail, Ingredient
-from config.config_manager import shared
+from src.config_manager import shared
 
 
 class DisplayController(DialogHandler):
