+++ src/maker.py
@@ -14,7 +14,7 @@ from src.display_controller import DP_CONTROLLER
 from src.service_handler import SERVICE_HANDLER
 from src.logger_handler import LoggerHandler
 
-from config.config_manager import shared
+from src.config_manager import shared
 
 
 LOG_HANDLER = LoggerHandler("maker_module", "production_logs")
