+++ src/service_handler.py
@@ -1,7 +1,7 @@
 import json
 from typing import Dict
 import requests
-from config.config_manager import ConfigManager
+from src.config_manager import ConfigManager
 from src.logger_handler import LoggerHandler
