+++ src_ui/setup_bottle_window.py
@@ -1,8 +1,8 @@
 from PyQt5.QtWidgets import QMainWindow
-from config.config_manager import ConfigManager
 
 from ui_elements.bottlewindow import Ui_Bottlewindow
 
+from src.config_manager import ConfigManager
 from src.bottles import set_fill_level_bars
 from src.database_commander import DB_COMMANDER
 from src.dialog_handler import UI_LANGUAGE
