+++ src_ui/setup_team_window.py
@@ -3,7 +3,7 @@ from PyQt5.QtGui import QIcon
 from PyQt5.QtWidgets import QDialog
 
 from ui_elements.teamselection import Ui_Teamselection
-from config.config_manager import ConfigManager, shared
+from src.config_manager import ConfigManager, shared
 from src.display_controller import DP_CONTROLLER
 from src.dialog_handler import UI_LANGUAGE
